move: svg파일 위치 변경 및 png파일 svg로 변경
